11726 - deleted RecipientResultsContainer.jsx
